Check if source name is resolvable in SIMBAD
simbad_name_resolvable in sources.py

So far, the method only returns true if the given source name and SIMBAD name are the exact same. I am working on a fuzzy search option to provide the user recommendations for an alternate SIMBAD resolvable name if original source name is not resolvable.